[UI] 과목 선택 버튼 수정 및 크기 조정, 파일 이름 수정
