+++ lib/main.dart
@@ -11,7 +11,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'pages/grading_page.dart';
 import 'pages/home_page.dart';
 import 'firebase_options.dart';
-import 'pages/select_exam_page.dart';
+import 'pages/select_full_exam_page.dart';
 import 'pages/test_page.dart';
 import 'class/grading_arguments.dart';
 import 'package:flutter/foundation.dart';
@@ -45,7 +45,7 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        initialRoute: '/fullExam',
+        initialRoute: '/selectPage',
         routes: {
           '/': (context) => const MyHomePage(),
           '/problemMake': (context) => const ProblemMake(),
@@ -58,7 +58,7 @@ class MyApp extends StatelessWidget {
               gradingArguments: ModalRoute.of(context)!.settings.arguments
                   as GradingArguments),
           '/weaknessExam': (context) => WeaknessExamPage(),
-          '/selectPage': (context) => SelectExamPage(),
+          '/selectPage': (context) => SelectFullExamPage(),
           '/unitExam': (context) => UnitExamPage(),
           '/unitExamGradingPage': (context) => UnitExamGradingPage(
               gradingArguments: ModalRoute.of(context)!.settings.arguments
